requirements.txt
matplotlib==3.5.0
numpy==1.21.2
pandas==1.3.5
Pillow==9.0.0
seaborn==0.11.2
streamlit==1.3.1
lxml<=4.7.1
